Implement clear favorites action from BoroughsSidebar.
Note: In the previous implementation Borough component checks whether is favorite when component is mounted.
The control of that behavior was responsibility of the internal logic of the component.
With the addition of clear favorites functionality,
Borough component should be able to reflect change from outside i.e. the parent component.
If we leave the internal state it should be calculated on each re-render to be able reflect the change, which is inefficient.
For that reason we just pass isFavorite prop, which is calulated in the parent
and avoid unnecessary calculation on each re-render of Borough component.